+++ src/main/java/writer/InputNumber.java
@@ -1,0 +1,10 @@
+package writer;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class InputNumber {
+    public String write(){
+        String userNumber = Console.readLine();
+        return userNumber;
+    }
+}
